Modified the XQilla class to use an enumeration to choose the language to parse.
Retrieve the default element namespace during static resoltion in a NodeTest
object. This makes it quicker to apply the node test, and easier to
read the URI it will try to match.

Added a return code to the XQTS test runner. Added a "known errors" file,
and (slightly modified) XQTS test suite zip, along with a script which runs
the test suite and returns a non-zero error code when it sees unexpected
errors in the tests. Users using the Makefile can use the make target
"test" to run the tests automatically.

--HG--
branch : xqilla