Reimplemented Timezone using a single MAPM recording the number of seconds offset.
Changed all ATGXXXOrDerived object to use a MAPM for their internal state,
rather than an ATDecimalOrDerived.

Added parseXXX() methods to the Items that use MAPMs for their internal
state. DB XML needs to be able to parse these objects easily for indexing.

Modified UnaryMinus to handle unary plus case as well, in order to give an
error when it's operand was incorrect.

--HG--
branch : xqilla