Abstracted the interfaces needed to write a new data model. All that is needed now is to implement the XQillaConfiguration class, and the classes it acts as a factory for.
Implemented a new data model called FastXDM. It places nodes in a
sequential array, and is much faster and more memory efficient than the DOM
based data model.

Moved the Xerces DOM based data model into the "xerces" directory, and
implemented the XercesConfiguration class to enable it's use.

Reimplemented node construction using a push-based event api. This avoids
the construction of intermediate results, and is therefore more efficient.

Added an AtomicObjectType enumeration parameter to AnyAtomicType::castAs().
This reduces the number of primitive type lookups that are done, and makes
casts to primitive types easier and more efficent.

Improved the algorithm in FunctionSum, FunctionAvg, FunctionMin and
FunctionMax. Removed AggregateFunction, as it is no longer used.

Created the Node::root() method, to allow an efficient implementation of
fn:root() to be written.

Optimised the fetching of the default Collation in GeneralComp.

Moved XQuery schema types creation to XQillaPlatformUtils.

Fixed a bug in the use of a locked grammar pool for a derived document
cache.

Renamed ASTNode::collapseTree() to createResult(), and
ASTNodeImpl::collapseTreeInternal() to createSequence().

Created a program to run the XMark benchmark tests. Work in progress.