Added XQQuery::execute(EventHandler*,DynamicContext*), which returns the query results to an EventHandler object. Used this API to send results directly to serialization in the XQTS runner and the "xqilla" command line executable.
Implemented XQUserFunction::XQFunctionEvaluator::generateEvents(), to
enable streaming of results from a user defined function.

Fixed XQQuantified by implementing XQQuantified::generateEvents(),
overriding the one it was inheriting from XQFLWOR.

Fixed NSFixupFilter to output the same prefix for an endElementEvent() as for
the startElementEvent().

Added an AnyAtomicType::AtomicObjectType parameter to
EventHandler::atomicItemEvent(), to better identify the type.