Converted common NodeTest objects that use an ItemType object to NodeTest objects without an ItemType object. This simplifies comparing different NodeTest objects, and uses a more efficient implementation.
Exposed methods on XQCastAs to perform the cast specified to an
AnyAtomicType and a string.

Modified MAPM so that some of it's common algorithms are thread safe, and
therefore reduced the use of the mutex to guard the library. The same MAPM
object should not now be used from multiple threads at the same time -
which means that the built-in MAPM constants should be avoided.

--HG--
branch : xqilla