Completely rewrote the lexer, using a different strategy. It is far simpler now, and uses far less states. All entities are now escaped by the lexer, as well as CDATA sections.
Rewrote a large part of the parser to use the new lexer tokens. Got rid of
things like merge_strings() in favour of lexing larger tokens. The
whitespace check on element content is now also performed solely in the
lexer. Simplified a number of the more complex rules, and reduced code
duplication.

Removed XQDebugHook, XQDebugCallback and XQRemoteDebugger.

Replaced VariableStore with a much simpler version, which is designed to be
sub-typed by the runtime objects.

Implemented closures and used them for "let" variables.

Reimplemented XQFLWOR as a set of simpler TupleNode objects that return a
TupleResult - basically a VariableStore with a next() method on it. These
include ForTuple, LetTuple, WhereTuple, and OrderByTuple, which are
contained inside an XQReturn object (implementing ASTNode).

Reimplemented XQQuantified using ForTuple objects.

Refactored or rewrote XQFunctionCall, XQTypeswitch, XQValidation,
XQUserFunction and UTransform.

--HG--
branch : xqilla