Implemented better static properties for node constructors.
Optimised fn:root(expr) to "expr" when it is statically known that the
argument is a document node.

Added the sourceforge logo to the top of every doxygen generated
documentation page. This will allow us to track page views better, since
the logo is used by sourceforge to calculate statistics.