Modified the XQTS test runner to run the XQuery Update test suite.
Backed out the XQuery Update syntax changes. They will need to be made at
some point, but the XQuery Update tests use the old syntax at the moment.

Fixed a bug in the test for updating expressions in XQIf.

Fixed the target node that was added to the pending update list for
attributes in UInsertBefore and UInsertAfter.

--HG--
branch : xqilla