Extracted schema validation from it's tight integration with the Xerces-C scanner classes into SchemaValidatorFilter, which allows us to add schema validation into an EventHandler pipeline.
Used SchemaValidatorFilter to reimplement XQValidate as part of the event
pipeline, so that it integrates seamlessly into node construction. This
means we avoid construction of a tree and subsequent serialization for the
argument to validate.

Used SchemaValidatorFilter to reimplement revalidation in the
XercesUpdateFactory, allowing a DOM sub-tree to be revalidated without
serializing and creating a parallel validated tree.