Added the xqilla:parse-html() function which uses HTML Tidy to parse HTML into XML. The function will not be present if TidyLib is not found.
Added fn:unparsed-text(), a function from XSLT 2.0 which can fetch a URL
and return it as a string.

Added xercesc_content_type.patch which can be applied against Xerces-C 2.8
to give access to the Content-Type header from HTTP connections, as well as
implementing redirects correctly. XQilla will detect if the patch has been
applied and use the functionality if present.

Fixed the XQTS test runner to output update results.

Fixed a couple of bugs in fulltext support.