Implemented the current draft of the XQC API.
Added MAPM::toDouble(), Numeric::asDouble() and Numeric::asInt().

Changed the function item invokation operator, so that the "=>" syntax
is not needed.

Fixed the SequenceType "document-node(schema-element(XX))" to work as
it should.

Fixed a number of bugs in the SchemaValidatorFilter and the XQuery
Update revalidation code.

Fixed the XercesConfiguration nodes to be able to be used as input
to secondary queries.

Fixed the DOM3 API to call fn:number() and fn:string() to get numbers
and strings.

Fixed the XQTS runner to only count update tests once, and to not copy
and re-write modified files to disk.

Bumped the project and library version numbers in preparation for
XQilla 2.2.